Added ability to download the latest database from Github
The program tries to download the latest database file from Github from https://raw.githubusercontent.com/SerjSX/pSearch/master/others/websitesdb

This is to prevent pushing releases for every site adding/removing.

FOR DEVELOPERS/CONTRIBUTORS: change testMode variable to True so it would connect to the local database, but make sure to turn it back to False if you are going to recommend it to me!